Don't allocate space for arguments and call frame if arguments aren't captured https://bugs.webkit.org/show_bug.cgi?id=95024
Reviewed by Phil Pizlo.

27% on v8-real-earley.

* runtime/JSActivation.h:
(JSC::JSActivation::registerOffset): The offset is zero if we're skipping
the arguments and call frame because "offset" means space reserved for
those things.

(JSC::JSActivation::tearOff): Don't copy the scope chain and callee. We
don't need them for anything, and we're no longer guaranteed to have
space for them.


Identifier: 112904@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126722 268f45cc-cd09-0410-ab3c-d52691b4dbfc