[WK2] Bugs in maintenance of internal state when user decides whether to grant notification permissions https://bugs.webkit.org/show_bug.cgi?id=95220 <rdar://problem/12189895>
Reviewed by Jessie Berlin.

A couple of the maps maintained by the request manager should have been cleaned up when the user decided on
whether to grant a website permission to post notifications.

Also, the web process' copy of the permissions was not updated appropriately. This meant that in the
permission callback, Notification.permission was not the same value as the permission value included as
the first parameter of the callback.

This first surfaced as part of the work to bring Mac support for web notifications. I have a test that
will check for regressions in this area, once all of that has been checked in (bug 77969).

* WebProcess/Notifications/NotificationPermissionRequestManager.cpp:
(WebKit::NotificationPermissionRequestManager::didReceiveNotificationPermissionDecision):

Identifier: 113080@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126903 268f45cc-cd09-0410-ab3c-d52691b4dbfc