Remove the now-unneeded invalidations in RenderTable::removeCaption https://bugs.webkit.org/show_bug.cgi?id=94889
Reviewed by Abhishek Arya.

Following bug 94842 and 95090, the invalidation code was pushed down to
RenderTableCaption. This made apparent that we did some invalidations that
were not needed.

Refactoring covered by existing tests.

* rendering/RenderTable.cpp:
(WebCore::RenderTable::removeCaption):
Removed setNeedsRecalcStyle() as it's unneeded now. It was probably needed back when
we didn't support multiple captions (see bug 58249) but the need was never documented
so you could wonder if it was really needed in the first place.

Identifier: 113310@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127139 268f45cc-cd09-0410-ab3c-d52691b4dbfc