[DRT] Make simulating a web click on a notification a queued task https://bugs.webkit.org/show_bug.cgi?id=95546 <rdar://problem/12214170>
Reviewed by Alexey Proskuryakov.

Making the web click happen asynchronously better mimics user interaction with the platform.

* DumpRenderTree/TestRunner.h: Add a flag that determines whether we have a pending notification click.
(TestRunner::hasPendingWebNotificationClick): Exposed so that dump() can check that the flag is
not set.
* DumpRenderTree/mac/DumpRenderTree.mm:
(dump): Assert that the flag is not set.
* DumpRenderTree/mac/TestRunnerMac.mm:
(TestRunner::simulateWebNotificationClick): Set the flag, then add a task to click the
notification. In the case where the block gets executed after the flag has been unset, we avoid
performing the click.

Identifier: 113457@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127298 268f45cc-cd09-0410-ab3c-d52691b4dbfc