ondisplay event handlers are not called https://bugs.webkit.org/show_bug.cgi?id=95263 <rdar://problem/12193359>
Reviewed by Alexey Proskuryakov.

Source/WebCore:

To support the legacy API, we conflate ondisplay and onshow.

* Modules/notifications/Notification.h:
(WebCore::Notification::ondisplay): Use display event as an alias for the show event.
(WebCore::Notification::setOndisplay):

LayoutTests:

Adjust the test to use the ondisplay handler, to fully test legacy API.

* http/tests/notifications/legacy/show-expected.txt: Verify that event listener for ondisplay is
the same as that for onshow.
* http/tests/notifications/legacy/show.html:

Identifier: 113480@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127321 268f45cc-cd09-0410-ab3c-d52691b4dbfc