[BlackBerry] ASSERT failure in RenderBlock::checkPositionedObjectsNeedLayout https://bugs.webkit.org/show_bug.cgi?id=95611
Reviewed by Antonio Gomes.
Patch by Jacky Jiang <zhajiang@rim.com>

PR:200127
Remove r118405 now as we already have a new fix webkit/43f8394b for
that issue and it is not reproducible anymore after the removal.
In this way, we can get rid of this ASSERT failure as well.
Thanks Antonio for the suggestion.

* Api/WebPage.cpp:
(BlackBerry::WebKit::WebPagePrivate::setViewportSize):

Identifier: 113484@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127326 268f45cc-cd09-0410-ab3c-d52691b4dbfc