[Qt] Automate the use of AllInOne files in the Qt build https://bugs.webkit.org/show_bug.cgi?id=95837
Reviewed by Tor Arne Vestbø.

Source/WebCore:

Replace the manual use of SVGAllInOne.cpp with automated use through
the ALL_IN_ONE_SOURCES variable. This adds a few more all-in-one sources
to the build and leaves out some that do not compile yet due to various
platform-dependant pre-processor macro issues.

* Target.pri:

Tools:

New feature file that implements the all-in-one file automation by
adding the requested all-in-one files to SOURCES and removing the files
it includes from SOURCES.

* qmake/mkspecs/features/use_all_in_one_files.prf: Added.

Identifier: 113711@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127576 268f45cc-cd09-0410-ab3c-d52691b4dbfc