[V8] V8DOMWindowShell does not need to be RefCounted https://bugs.webkit.org/show_bug.cgi?id=96192
Reviewed by Kentaro Hara.

V8DOMWindowShell holds a raw Frame pointer, which would point to
unallocated memory if any object that wasn't owned by the frame took a
reference to a V8DOMWindowShell. Thankfully, the only object that takes
a reference to the V8DOMWindowShell is ScriptController, which is owned
by the Frame.

This patch stops V8DOMWindowShell from being RefCounted so that no
other objects are tempted to take a reference to it in the future.

* bindings/v8/ScriptController.h:
(ScriptController):
* bindings/v8/V8DOMWindowShell.cpp:
(WebCore::V8DOMWindowShell::create):
* bindings/v8/V8DOMWindowShell.h:
(V8DOMWindowShell):


Identifier: 114075@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127977 268f45cc-cd09-0410-ab3c-d52691b4dbfc