[EFL] Wrong rendering results for square-button-appearance.html https://bugs.webkit.org/show_bug.cgi?id=85493
Patch by KyungTae Kim <ktf.kim@samsung.com> on 2012-09-09
Reviewed by Gyuyoung Kim.

Source/WebCore:

A square-button is an element that has '-webkit-appearance: square-button' style.

If it is a block element,
the RenderBlock::recomputeLogicalWidth recalculate the width from the padding values.

But the recalculated width became wrong because the RenderThemeEfl::adjustSizeConstraints
called from RenderThemeEfl::adjustButtonStyle adds padding values for a button.

So, the adjustSizeConstraints must not be called for a square-button(SquareButtonPart).

Test : fast/css/square-button-appearance.html

* platform/efl/RenderThemeEfl.cpp:
(WebCore::RenderThemeEfl::adjustButtonStyle): Call adjustSizeConstrains only for PushButtonPart and ButtonPart

LayoutTests:

Remove fast/css/square-button-appearance.html from TestExpectations because it became PASS.

* platform/efl/TestExpectations:

Identifier: 114108@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128013 268f45cc-cd09-0410-ab3c-d52691b4dbfc