[WK2] Make WK2 C API headers compile in the Qt build on Windows https://bugs.webkit.org/show_bug.cgi?id=96246
Patch by Simon Hausmann <simon.hausmann@nokia.com> on 2012-09-10
Reviewed by Kenneth Rohde Christiansen.

WKBase.h and WKAPICast.h include files that are specific to the Safari/Windows
port if WIN32 or _WIN32 is defined. When building WebKit2 on Windows with Qt, we
do not want to include these files, as they define an API that is different to ours
(the view aspect differs). Therefore guard the inclusion with !defined(BUILDING_QT__)

* Shared/API/c/WKBase.h:
* UIProcess/API/C/WKAPICast.h:

Identifier: 114128@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128033 268f45cc-cd09-0410-ab3c-d52691b4dbfc