Source/WebCore: Clarify the cause of console warnings generated by "cross-origin" access to sandboxed iframes. https://bugs.webkit.org/show_bug.cgi?id=64079
Patch by Mike West <mkwst@chromium.org> on 2012-09-10
Reviewed by Adam Barth.

The error message generated when accessing cross-origin content is a bit
too generic at the moment, which is misleading when the "cross-origin"
nature of an access attempt isn't visible in the URLs that the error
displays. Sandboxed iframes, for example, are put into unique origins by
default, meaning that all access in or out are subject to cross-origin
restrictions, even if the resources would otherwise be same-origin.

This patch improces the error message to explicitly point to sandboxing
as the core of cross-origin requests when relevant. It adds a single new
test to check access from parents to children, and relies on existing
tests for access in the other direction.

Test: http/tests/security/sandboxed-iframe-blocks-access-from-parent.html

* page/DOMWindow.cpp:
(WebCore::DOMWindow::crossDomainAccessErrorMessage):
    Check 'isSandboxed(SandboxOrigin)' against the document and
    activeWindow when generating the error message. If one or the other
    is sandboxed, make that clear in the console warning.

LayoutTests: Console warnings generated by script access to sandboxed iframes should make the violation clear.
https://bugs.webkit.org/show_bug.cgi?id=64079

Patch by Mike West <mkwst@chromium.org> on 2012-09-10
Reviewed by Adam Barth.

* http/tests/security/resources/blank.html: Added.
    Adding a blank resource for a same-origin access test.
* http/tests/security/sandboxed-iframe-blocks-access-from-parent-expected.txt: Added.
* http/tests/security/sandboxed-iframe-blocks-access-from-parent.html: Added.
    Parents shouldn't be able to access properties (like location) of
    their sandboxed child frames.
* http/tests/security/sandboxed-iframe-modify-self-expected.txt:
* http/tests/security/sandboxed-iframe-origin-add-expected.txt:
* http/tests/security/sandboxed-iframe-origin-remove-expected.txt:
* http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt:
* http/tests/security/xss-DENIED-sandboxed-iframe-expected.txt:
* platform/chromium/http/tests/security/sandboxed-iframe-modify-self-expected.txt:
* platform/chromium/http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt:
    Updating existing error messages.

Identifier: 114163@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128070 268f45cc-cd09-0410-ab3c-d52691b4dbfc