REGRESSION (r127936): Multiple accessibility tests failing on Lion https://bugs.webkit.org/show_bug.cgi?id=96173
Reviewed by Chris Fleizach.

This test (update-children-when-aria-role-changes.html) assumes that
an element with an aria-label is ignored, which worked before, but it's
not ignored anymore after r127936. One of the comments in
AccessibilityRenderObject.cpp says: "if this element has aria attributes on it,
it should not be ignored," so I think the test should be updated to not use
aria-label on an element we want to be ignored.

* platform/mac/Skipped:
* platform/mac/accessibility/update-children-when-aria-role-changes.html:


Identifier: 114165@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128072 268f45cc-cd09-0410-ab3c-d52691b4dbfc