[chromium, android] Reloading a page with a different user agent can cause the page to be zoomed in https://bugs.webkit.org/show_bug.cgi?id=90222
Patch by Dan Alcantara <dfalcantara@chromium.org> on 2012-09-10
Reviewed by Adam Barth.

When reloading a page with an overridden URL, the page's scroll and zoom
state will be restored once the reload is complete.  This is problematic
in situations when switching between mobile and desktop versions of the
same site because it will zoom back in on a random part of alternate page.
This CL just avoids that situation by resetting the zoom and scale whenever
this type of reload occurs.

Chromium half: https://chromiumcodereview.appspot.com/10889019

* public/WebView.h:
(WebView):
* src/WebViewImpl.cpp:
(WebKit::WebViewImpl::resetScrollAndScaleState):
(WebKit):
* src/WebViewImpl.h:
(WebViewImpl):

Identifier: 114226@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128133 268f45cc-cd09-0410-ab3c-d52691b4dbfc